Add files via upload
The MANAV SEVA.COM (hospital management system) is a web-based application. It is used to manage the field of a hospital such as doctors, patients, ambulance service, pharmacy and appointments. 
The system would help the hospital to continue their services in a quick and cost-effective manner. It is also helpful for patient parties to boo bed and ambulances quickly. 
The salient features of this system are:
1. Keeps record of each and every doctor, whether he or she is available or not.
2. Stores all the medicine details along with their manufacture and expiry date.
3. Allows hospital to admit any patient instantly through an online form.
4. Provide details of all the admitted patients.
5. Display all the appointments for the hospital. It also allows people to fix their appointments at any hospital.
6. Provides access to book ambulance of any hospital by the patient parties .

The above files contain all the html, css, php codes and pictures.